requirements.txt
pyserial==3.5
aiohttp==3.8.4
aiohttp-jinja2==1.5
aioopenssl==0.6.0
aiosasl==0.5.0
aiosignal==1.3.1
aioxmpp==0.13.3
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.0.3
jinja2-time==0.2.0
lxml==4.9.2
MarkupSafe==2.1.3
multidict==5.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2022.1
singletonify==0.2.4
six==1.16.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
spade==3.3.0
timeago==1.0.16
tzdata==2023.3
tzlocal==5.0.1
yarl==1.9.2
paho